NamedCommands fixes, command updates
